+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [10.0.8] 2025-06-09
+## [10.1.1] 2025-06-09
+- Fixed copying of an item when a single item is in the sidebar tree.
+- 
+## [10.1.0] 2025-05-29
+### Added
+- Enabled combining multiple displays and codes in a unit lookup.
+- Added a variable extension at the questionnaire level.
 ### Fixed
-- Fixed copy of an item when a single item is in the sidebar tree.
+- Fixed the no_match status when switching between Pick initial value and other value methods.
+- Fixed the empty answer option check for all data types except coding.
 
 ## [10.0.7] 2025-05-19
 ### Fixed